% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mod.R
\name{format_cis}
\alias{format_cis}
\title{Format confidence intervals}
\usage{
format_cis(cis, digits = 2)
}
\arguments{
\item{cis}{(data.frame): CIs with low in first column high in second}

\item{digits}{(integer): precision of result}
}
\description{
Returns vector of strings with formatted confidence intervals
}
