% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mod.R
\name{format_num}
\alias{format_num}
\title{Format number}
\usage{
format_num(num, digits = 2)
}
\arguments{
\item{num}{(vector): vectors of numbers to format}

\item{digits}{(integer): precision of result}
}
\value{
(vector): strings of fixed width formatted numbers
}
\description{
Format number
}
