% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mod.R
\name{get_sci}
\alias{get_sci}
\title{Convert number to scientific notation}
\usage{
get_sci(num, digits = 2)
}
\arguments{
\item{num}{(numeric): number to convert}

\item{digits}{(integer): precision of result}
}
\description{
Returns a string version of number in scientific notation.
}
