% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mod.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Checks if a number is whole}
\usage{
is_wholenumber(num, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{num}{(numeric): number to check}

\item{tol}{(numeric): how close the number needs to be to whole}
}
\description{
Returns whether a number is whole
}
