% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mod.R
\name{pretty_mod}
\alias{pretty_mod}
\title{Pretty format results from models}
\usage{
pretty_mod(
  mod,
  type = "binomial",
  digits = 2,
  flex_caption = NULL,
  expo = TRUE,
  skip = 1
)
}
\arguments{
\item{mod}{(object): a model object}

\item{type}{(character): the type of model (i.e. 'binomial', 'negbin', 'surv', 'ordinal', 'lm')}

\item{digits}{(integer): precision of result}

\item{flex_caption}{(character): caption for flex table}

\item{expo}{(Boolean): take exponential function of estimate}
}
\value{
(list): data.frame of results or df and flextable
}
\description{
Pretty format results from models
}
