% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mod.R
\name{pretty_mod}
\alias{pretty_mod}
\title{Pretty format results from models}
\usage{
pretty_mod(mod, type = "binomial", digits = 2, flex_caption = NULL)
}
\arguments{
\item{type}{(character): the type of model}

\item{digits}{(integer): precision of result}

\item{flex_caption}{(character): caption for flex table}

\item{mod:}{a model object}
}
\description{
Returns list with data.frame or data.frame and flextable
}
